Drag and Drop: refactor, add more styling
[#176192238]
[#176013709]
