+++ src/drag-and-drop/components/draggable-item-wrapper.scss
@@ -1,0 +1,15 @@
+@import "../../shared/styles/helpers";
+
+.draggableItemWrapper {
+  position: absolute;
+  cursor: move;
+  &:hover {
+    $border-width: 4px;
+    border: $border-width solid yellow;
+    transform: translate(-$border-width, -$border-width);
+  }
+
+  img {
+    pointer-events: none;
+  }
+}
